entry_task/requirements.txt
Django==1.11.20
python-memcached
crypto
jsonschema
python-dateutil
mysql-connector-python